Revert "chore : create atom components"
