+++ src/App.tsx
@@ -1,15 +1,20 @@
-import { ThemeProvider } from 'styled-components';
+import styled from 'styled-components';
 import Router from './router/Router';
 import GlobalStyle from './styles/globalStyles';
-import theme from './styles/theme';
 
 function App() {
   return (
-    <ThemeProvider theme={theme}>
+    <Container>
       <GlobalStyle/>
       <Router/>
-    </ThemeProvider>
+    </Container>
   );
 }
+
+const Container = styled.div`
+  display: flex;
+  justify-content: center;
+  width: 100%;
+`
 
 export default App;
