--- src/components/Atoms/Buttons/TranslationButton.tsx
@@ -1,19 +1,0 @@
-import { useState } from "react";
-import styled from "styled-components";
-
-const StyledButtonWrap = styled.button`
-
-`
-
-const TranslationButton = () => {
-  const [clicked, setClick] = useState<boolean>(false)
-  return (
-    <>
-      <StyledButtonWrap onClick={() => setClick(!clicked)}>
-        {clicked ? "eng" : "kor"} 
-      </StyledButtonWrap>
-    </>
-  )
-}
-
-export default TranslationButton;
