+++ src/components/Atoms/ModeButton.tsx
@@ -1,0 +1,5 @@
+const ModeButton = () => {
+  return (
+    <></>
+  )
+}
