+++ src/components/Pages/AboutMe.tsx
@@ -1,10 +1,6 @@
-import { Header } from "../Molecules/Header";
-
 const AboutMe = () => {
   return (
-    <>
-      <Header />
-    </>
+    <></>
   )
 }
