+++ src/components/Pages/Contact.tsx
@@ -1,10 +1,6 @@
-import { MainHeader } from "../Molecules/Header";
-
 const Contact = () => {
   return (
-    <>
-      <MainHeader/>
-    </>
+    <></>
   )
 }
