+++ src/components/Pages/Project.tsx
@@ -1,10 +1,6 @@
-import { Header } from "../Molecules/Header";
-
 const Project = () => {
   return (
-    <>
-      <Header/>
-    </>
+    <></>
   )
 }
